src/requirements.txt
absl-py==1.3.0
altair==4.2.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.4
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
gensim==4.2.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
inflect==6.0.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
protobuf==3.19.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
selenium==4.7.2
semver==2.13.0
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
streamlit==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-macos==2.11.0
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.11.0
